Add option to show security attribute and improve extended support
Add a command line option -Z/--context to show the security context of
objects, similar to ls(1).

Show the actual extended attribute values on -@/--extended, instead of
just their length.

In case of a symbolic link, show the extended attributes of the symbolic
link itself, not the target. This matches the behavior of ls(1) and is
more intuitive.

Closes: #254