Make `-t` work like ls when used with no args
Currently exa can't work as a drop-in `ls` replacement due to `-t`
having a very different behaviour.

However `-t` currently only works when used with a value, and is invalid
when used without any value.

This patch changes the behaviour of exa when `-t` is passed with no
value to behave the same as ls. When used with a value, it continues
behaving as in previous versions of `exa`.

Fixes: https://github.com/ogham/exa/issues/519