Symlink dereferencing: add flag and propagate it.
Add the `-X`/`--dereference` flags which are used to enable
dereferencing through symbolic links, similar to ls's `-L`. Added the
`link_target_recurse` method in order to recursively call the existing
`link_target` method to get the final target at the end of the chain of
links. In order to appropriately display the final output and play
nicely with existing functionality, a `deref_links` variable has been
added to a few places such that it can be propagated until the time it
is needed during rendering. Following commits will use this information
to display dereferenced information.